@@ -10,7 +10,7 @@
     "seed": "ts-node scripts/seed.ts",
     "db:migrate-save": "prisma migrate dev",
     "db:migrate-up": "prisma migrate deploy",
-    "db:clean": "ts-node scripts/clean.ts",
+    "db:clean": "prisma migrate reset",
     "db:init": "run-s \"db:migrate-save -- --name 'initial version'\" db:migrate-up seed",
     "prisma:generate": "prisma generate",
     "docker:dev": "docker-compose -f docker-compose.dev.yml up -d",
@@ -19,57 +19,58 @@
     "compose:down": "docker-compose down --volumes"
   },
   "dependencies": {
-    "@nestjs/common": "8.4.7",
-    "@nestjs/config": "1.1.5",
-    "@nestjs/core": "8.4.7",
-    "@nestjs/graphql": "9.1.2",
-    "@nestjs/platform-express": "8.4.7",
-    "@nestjs/serve-static": "2.2.2",
-    "@nestjs/swagger": "5.1.5",
+    "@apollo/server": "^4.9.4",
+    "@nestjs/apollo": "12.0.9",
+    "@nestjs/common": "10.2.7",
+    "@nestjs/config": "3.1.1",
+    "@nestjs/core": "10.2.7",
+    "@nestjs/graphql": "12.0.9",
+    "@nestjs/jwt": "^10.1.1",
+    "@nestjs/passport": "^10.0.2",
+    "@nestjs/platform-express": "10.2.7",
+    "@nestjs/serve-static": "4.0.0",
+    "@nestjs/swagger": "7.1.13",
+    "@prisma/client": "^5.4.2",
     "@types/bcrypt": "5.0.0",
-    "@prisma/client": "4.6.1",
-    "apollo-server-express": "3.6.1",
-    "bcrypt": "5.0.1",
+    "bcrypt": "5.1.1",
     "class-transformer": "0.5.1",
     "class-validator": "0.14.0",
-    "dotenv": "^16.1.4",
-    "graphql": "15.7.2",
+    "dotenv": "16.3.1",
+    "graphql": "^16.8.1",
     "graphql-type-json": "0.3.2",
-    "nest-access-control": "2.0.3",
-    "nest-morgan": "1.0.1",
     "npm-run-all": "4.1.5",
-    "reflect-metadata": "0.1.13",
-    "swagger-ui-express": "4.3.0",
-    "ts-node": "10.9.1",
-    "type-fest": "0.13.1",
-    "validator": "^13.9.0",
-    "@nestjs/jwt": "^10.0.2",
-    "@nestjs/passport": "^9.0.0",
     "passport": "0.6.0",
     "passport-http": "0.3.0",
-    "passport-jwt": "4.0.1"
+    "passport-jwt": "4.0.1",
+    "reflect-metadata": "0.1.13",
+    "ts-node": "10.9.2",
+    "type-fest": "2.19.0",
+    "validator": "13.11.0"
   },
   "devDependencies": {
-    "@nestjs/cli": "8.2.5",
-    "@nestjs/testing": "8.4.7",
-    "@types/express": "4.17.9",
-    "@types/graphql-type-json": "0.3.2",
-    "@types/jest": "26.0.19",
+    "@nestjs/cli": "^10.1.18",
+    "@nestjs/testing": "^10.2.7",
+    "@types/express": "^4.17.19",
+    "@types/graphql-type-json": "0.3.3",
+    "@types/jest": "^29.5.5",
     "@types/normalize-path": "3.0.0",
-    "@types/supertest": "2.0.11",
-    "@types/validator": "^13.7.15",
-    "jest": "27.0.6",
-    "jest-mock-extended": "^2.0.4",
-    "prisma": "4.6.1",
-    "supertest": "4.0.2",
-    "ts-jest": "27.0.3",
-    "typescript": "4.3.5",
     "@types/passport-http": "0.3.9",
-    "@types/passport-jwt": "3.0.8"
+    "@types/passport-jwt": "3.0.10",
+    "@types/supertest": "^2.0.14",
+    "@types/validator": "^13.11.2",
+    "jest": "^29.7.0",
+    "jest-mock-extended": "^3.0.5",
+    "prisma": "^5.4.2",
+    "supertest": "^6.3.3",
+    "ts-jest": "^29.1.1",
+    "typescript": "^5.4.3"
   },
   "jest": {
     "preset": "ts-jest",
     "testEnvironment": "node",
+    "moduleNameMapper": {
+      "@app/custom-validators": "<rootDir>/src/validators"
+    },
     "modulePathIgnorePatterns": ["<rootDir>/dist/"]
   }
 }
