@@ -19,13 +19,12 @@ export const AddressCreate = (props: CreateProps): React.ReactElement => {
         <TextInput label="Address 1" source="address_1" />
         <TextInput label="Address 2" source="address_2" />
         <TextInput label="City" source="city" />
-        <ReferenceArrayInput
-          source="customers"
-          reference="Customer"
-          parse={(value: any) => value && value.map((v: any) => ({ id: v }))}
-          format={(value: any) => value && value.map((v: any) => v.id)}
-        >
-          <SelectArrayInput optionText={CustomerTitle} />
+        <ReferenceArrayInput source="customers" reference="Customer">
+          <SelectArrayInput
+            optionText={CustomerTitle}
+            parse={(value: any) => value && value.map((v: any) => ({ id: v }))}
+            format={(value: any) => value && value.map((v: any) => v.id)}
+          />
         </ReferenceArrayInput>
         <TextInput label="State" source="state" />
         <NumberInput step={1} label="Zip" source="zip" />
