@@ -6,12 +6,11 @@ export const AddressList = (props: ListProps): React.ReactElement => {
   return (
     <List
       {...props}
-      bulkActionButtons={false}
       title={"Addresses"}
       perPage={50}
       pagination={<Pagination />}
     >
-      <Datagrid rowClick="show">
+      <Datagrid rowClick="show" bulkActionButtons={false}>
         <TextField label="Address 1" source="address_1" />
         <TextField label="Address 2" source="address_2" />
         <TextField label="City" source="city" />
@@ -19,7 +18,7 @@ export const AddressList = (props: ListProps): React.ReactElement => {
         <TextField label="ID" source="id" />
         <TextField label="State" source="state" />
         <DateField source="updatedAt" label="Updated At" />
-        <TextField label="Zip" source="zip" />
+        <TextField label="Zip" source="zip" />{" "}
       </Datagrid>
     </List>
   );
