@@ -13,14 +13,8 @@ import { PRODUCT_TITLE_FIELD } from "../product/ProductTitle";
 
 export const OrderList = (props: ListProps): React.ReactElement => {
   return (
-    <List
-      {...props}
-      bulkActionButtons={false}
-      title={"Orders"}
-      perPage={50}
-      pagination={<Pagination />}
-    >
-      <Datagrid rowClick="show">
+    <List {...props} title={"Orders"} perPage={50} pagination={<Pagination />}>
+      <Datagrid rowClick="show" bulkActionButtons={false}>
         <DateField source="createdAt" label="Created At" />
         <ReferenceField
           label="Customer"
@@ -36,7 +30,7 @@ export const OrderList = (props: ListProps): React.ReactElement => {
         </ReferenceField>
         <TextField label="Quantity" source="quantity" />
         <TextField label="Total Price" source="totalPrice" />
-        <DateField source="updatedAt" label="Updated At" />
+        <DateField source="updatedAt" label="Updated At" />{" "}
       </Datagrid>
     </List>
   );
