@@ -20,21 +20,19 @@ export const ProductEdit = (props: EditProps): React.ReactElement => {
         <TextInput label="Description" multiline source="description" />
         <NumberInput label="Item Price" source="itemPrice" />
         <TextInput label="Name" source="name" />
-        <ReferenceArrayInput
-          source="orders"
-          reference="Order"
-          parse={(value: any) => value && value.map((v: any) => ({ id: v }))}
-          format={(value: any) => value && value.map((v: any) => v.id)}
-        >
-          <SelectArrayInput optionText={OrderTitle} />
+        <ReferenceArrayInput source="orders" reference="Order">
+          <SelectArrayInput
+            optionText={OrderTitle}
+            parse={(value: any) => value && value.map((v: any) => ({ id: v }))}
+            format={(value: any) => value && value.map((v: any) => v.id)}
+          />
         </ReferenceArrayInput>
-        <ReferenceArrayInput
-          source="variasi"
-          reference="Variasi"
-          parse={(value: any) => value && value.map((v: any) => ({ id: v }))}
-          format={(value: any) => value && value.map((v: any) => v.id)}
-        >
-          <SelectArrayInput optionText={VariasiTitle} />
+        <ReferenceArrayInput source="variasi" reference="Variasi">
+          <SelectArrayInput
+            optionText={VariasiTitle}
+            parse={(value: any) => value && value.map((v: any) => ({ id: v }))}
+            format={(value: any) => value && value.map((v: any) => v.id)}
+          />
         </ReferenceArrayInput>
       </SimpleForm>
     </Edit>
