@@ -25,7 +25,7 @@ export const ProductShow = (props: ShowProps): React.ReactElement => {
         <TextField label="Name" source="name" />
         <DateField source="updatedAt" label="Updated At" />
         <ReferenceManyField reference="Order" target="productId" label="Orders">
-          <Datagrid rowClick="show">
+          <Datagrid rowClick="show" bulkActionButtons={false}>
             <DateField source="createdAt" label="Created At" />
             <ReferenceField
               label="Customer"
