@@ -16,41 +16,24 @@ import * as errors from "../../errors";
 import { Request } from "express";
 import { plainToClass } from "class-transformer";
 import { ApiNestedQuery } from "../../decorators/api-nested-query.decorator";
-import * as nestAccessControl from "nest-access-control";
-import * as defaultAuthGuard from "../../auth/defaultAuth.guard";
 import { AddressService } from "../address.service";
-import { AclValidateRequestInterceptor } from "../../interceptors/aclValidateRequest.interceptor";
-import { AclFilterResponseInterceptor } from "../../interceptors/aclFilterResponse.interceptor";
 import { AddressCreateInput } from "./AddressCreateInput";
-import { AddressWhereInput } from "./AddressWhereInput";
-import { AddressWhereUniqueInput } from "./AddressWhereUniqueInput";
-import { AddressFindManyArgs } from "./AddressFindManyArgs";
-import { AddressUpdateInput } from "./AddressUpdateInput";
 import { Address } from "./Address";
+import { AddressFindManyArgs } from "./AddressFindManyArgs";
+import { AddressWhereUniqueInput } from "./AddressWhereUniqueInput";
+import { AddressUpdateInput } from "./AddressUpdateInput";
 import { CustomerFindManyArgs } from "../../customer/base/CustomerFindManyArgs";
 import { Customer } from "../../customer/base/Customer";
 import { CustomerWhereUniqueInput } from "../../customer/base/CustomerWhereUniqueInput";
 
-@swagger.ApiBasicAuth()
-@common.UseGuards(defaultAuthGuard.DefaultAuthGuard, nestAccessControl.ACGuard)
 export class AddressControllerBase {
-  constructor(
-    protected readonly service: AddressService,
-    protected readonly rolesBuilder: nestAccessControl.RolesBuilder
-  ) {}
-  @common.UseInterceptors(AclValidateRequestInterceptor)
+  constructor(protected readonly service: AddressService) {}
   @common.Post()
   @swagger.ApiCreatedResponse({ type: Address })
-  @nestAccessControl.UseRoles({
-    resource: "Address",
-    action: "create",
-    possession: "any",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async create(@common.Body() data: AddressCreateInput): Promise<Address> {
-    return await this.service.create({
+  async createAddress(
+    @common.Body() data: AddressCreateInput
+  ): Promise<Address> {
+    return await this.service.createAddress({
       data: data,
       select: {
         address_1: true,
@@ -65,21 +48,12 @@ export class AddressControllerBase {
     });
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @common.Get()
   @swagger.ApiOkResponse({ type: [Address] })
   @ApiNestedQuery(AddressFindManyArgs)
-  @nestAccessControl.UseRoles({
-    resource: "Address",
-    action: "read",
-    possession: "any",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async findMany(@common.Req() request: Request): Promise<Address[]> {
+  async addresses(@common.Req() request: Request): Promise<Address[]> {
     const args = plainToClass(AddressFindManyArgs, request.query);
-    return this.service.findMany({
+    return this.service.addresses({
       ...args,
       select: {
         address_1: true,
@@ -94,22 +68,13 @@ export class AddressControllerBase {
     });
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @common.Get("/:id")
   @swagger.ApiOkResponse({ type: Address })
   @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
-  @nestAccessControl.UseRoles({
-    resource: "Address",
-    action: "read",
-    possession: "own",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async findOne(
+  async address(
     @common.Param() params: AddressWhereUniqueInput
   ): Promise<Address | null> {
-    const result = await this.service.findOne({
+    const result = await this.service.address({
       where: params,
       select: {
         address_1: true,
@@ -130,24 +95,15 @@ export class AddressControllerBase {
     return result;
   }
 
-  @common.UseInterceptors(AclValidateRequestInterceptor)
   @common.Patch("/:id")
   @swagger.ApiOkResponse({ type: Address })
   @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
-  @nestAccessControl.UseRoles({
-    resource: "Address",
-    action: "update",
-    possession: "any",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async update(
+  async updateAddress(
     @common.Param() params: AddressWhereUniqueInput,
     @common.Body() data: AddressUpdateInput
   ): Promise<Address | null> {
     try {
-      return await this.service.update({
+      return await this.service.updateAddress({
         where: params,
         data: data,
         select: {
@@ -174,19 +130,11 @@ export class AddressControllerBase {
   @common.Delete("/:id")
   @swagger.ApiOkResponse({ type: Address })
   @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
-  @nestAccessControl.UseRoles({
-    resource: "Address",
-    action: "delete",
-    possession: "any",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async delete(
+  async deleteAddress(
     @common.Param() params: AddressWhereUniqueInput
   ): Promise<Address | null> {
     try {
-      return await this.service.delete({
+      return await this.service.deleteAddress({
         where: params,
         select: {
           address_1: true,
@@ -209,15 +157,9 @@ export class AddressControllerBase {
     }
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @common.Get("/:id/customers")
   @ApiNestedQuery(CustomerFindManyArgs)
-  @nestAccessControl.UseRoles({
-    resource: "Customer",
-    action: "read",
-    possession: "any",
-  })
-  async findManyCustomers(
+  async findCustomers(
     @common.Req() request: Request,
     @common.Param() params: AddressWhereUniqueInput
   ): Promise<Customer[]> {
@@ -249,11 +191,6 @@ export class AddressControllerBase {
   }
 
   @common.Post("/:id/customers")
-  @nestAccessControl.UseRoles({
-    resource: "Address",
-    action: "update",
-    possession: "any",
-  })
   async connectCustomers(
     @common.Param() params: AddressWhereUniqueInput,
     @common.Body() body: CustomerWhereUniqueInput[]
@@ -263,7 +200,7 @@ export class AddressControllerBase {
         connect: body,
       },
     };
-    await this.service.update({
+    await this.service.updateAddress({
       where: params,
       data,
       select: { id: true },
@@ -271,11 +208,6 @@ export class AddressControllerBase {
   }
 
   @common.Patch("/:id/customers")
-  @nestAccessControl.UseRoles({
-    resource: "Address",
-    action: "update",
-    possession: "any",
-  })
   async updateCustomers(
     @common.Param() params: AddressWhereUniqueInput,
     @common.Body() body: CustomerWhereUniqueInput[]
@@ -285,7 +217,7 @@ export class AddressControllerBase {
         set: body,
       },
     };
-    await this.service.update({
+    await this.service.updateAddress({
       where: params,
       data,
       select: { id: true },
@@ -293,11 +225,6 @@ export class AddressControllerBase {
   }
 
   @common.Delete("/:id/customers")
-  @nestAccessControl.UseRoles({
-    resource: "Address",
-    action: "update",
-    possession: "any",
-  })
   async disconnectCustomers(
     @common.Param() params: AddressWhereUniqueInput,
     @common.Body() body: CustomerWhereUniqueInput[]
@@ -307,7 +234,7 @@ export class AddressControllerBase {
         disconnect: body,
       },
     };
-    await this.service.update({
+    await this.service.updateAddress({
       where: params,
       data,
       select: { id: true },
