@@ -10,47 +10,41 @@ https://docs.amplication.com/how-to/custom-code
 ------------------------------------------------------------------------------
   */
 import { PrismaService } from "../../prisma/prisma.service";
-import { Prisma, Address, Customer } from "@prisma/client";
+import {
+  Prisma,
+  Address as PrismaAddress,
+  Customer as PrismaCustomer,
+} from "@prisma/client";
 
 export class AddressServiceBase {
   constructor(protected readonly prisma: PrismaService) {}
 
-  async count<T extends Prisma.AddressCountArgs>(
-    args: Prisma.SelectSubset<T, Prisma.AddressCountArgs>
-  ): Promise<number> {
+  async count(args: Omit<Prisma.AddressCountArgs, "select">): Promise<number> {
     return this.prisma.address.count(args);
   }
 
-  async findMany<T extends Prisma.AddressFindManyArgs>(
-    args: Prisma.SelectSubset<T, Prisma.AddressFindManyArgs>
-  ): Promise<Address[]> {
+  async addresses(args: Prisma.AddressFindManyArgs): Promise<PrismaAddress[]> {
     return this.prisma.address.findMany(args);
   }
-  async findOne<T extends Prisma.AddressFindUniqueArgs>(
-    args: Prisma.SelectSubset<T, Prisma.AddressFindUniqueArgs>
-  ): Promise<Address | null> {
+  async address(
+    args: Prisma.AddressFindUniqueArgs
+  ): Promise<PrismaAddress | null> {
     return this.prisma.address.findUnique(args);
   }
-  async create<T extends Prisma.AddressCreateArgs>(
-    args: Prisma.SelectSubset<T, Prisma.AddressCreateArgs>
-  ): Promise<Address> {
-    return this.prisma.address.create<T>(args);
+  async createAddress(args: Prisma.AddressCreateArgs): Promise<PrismaAddress> {
+    return this.prisma.address.create(args);
   }
-  async update<T extends Prisma.AddressUpdateArgs>(
-    args: Prisma.SelectSubset<T, Prisma.AddressUpdateArgs>
-  ): Promise<Address> {
-    return this.prisma.address.update<T>(args);
+  async updateAddress(args: Prisma.AddressUpdateArgs): Promise<PrismaAddress> {
+    return this.prisma.address.update(args);
   }
-  async delete<T extends Prisma.AddressDeleteArgs>(
-    args: Prisma.SelectSubset<T, Prisma.AddressDeleteArgs>
-  ): Promise<Address> {
+  async deleteAddress(args: Prisma.AddressDeleteArgs): Promise<PrismaAddress> {
     return this.prisma.address.delete(args);
   }
 
   async findCustomers(
     parentId: string,
     args: Prisma.CustomerFindManyArgs
-  ): Promise<Customer[]> {
+  ): Promise<PrismaCustomer[]> {
     return this.prisma.address
       .findUniqueOrThrow({
         where: { id: parentId },
