@@ -10,41 +10,25 @@ https://docs.amplication.com/how-to/custom-code
 ------------------------------------------------------------------------------
   */
 import * as graphql from "@nestjs/graphql";
-import * as apollo from "apollo-server-express";
+import { GraphQLError } from "graphql";
 import { isRecordNotFoundError } from "../../prisma.util";
 import { MetaQueryPayload } from "../../util/MetaQueryPayload";
-import * as nestAccessControl from "nest-access-control";
-import * as gqlACGuard from "../../auth/gqlAC.guard";
-import { GqlDefaultAuthGuard } from "../../auth/gqlDefaultAuth.guard";
-import * as common from "@nestjs/common";
-import { AclFilterResponseInterceptor } from "../../interceptors/aclFilterResponse.interceptor";
-import { AclValidateRequestInterceptor } from "../../interceptors/aclValidateRequest.interceptor";
-import { CreateProductArgs } from "./CreateProductArgs";
-import { UpdateProductArgs } from "./UpdateProductArgs";
-import { DeleteProductArgs } from "./DeleteProductArgs";
+import { Product } from "./Product";
 import { ProductCountArgs } from "./ProductCountArgs";
 import { ProductFindManyArgs } from "./ProductFindManyArgs";
 import { ProductFindUniqueArgs } from "./ProductFindUniqueArgs";
-import { Product } from "./Product";
+import { CreateProductArgs } from "./CreateProductArgs";
+import { UpdateProductArgs } from "./UpdateProductArgs";
+import { DeleteProductArgs } from "./DeleteProductArgs";
 import { OrderFindManyArgs } from "../../order/base/OrderFindManyArgs";
 import { Order } from "../../order/base/Order";
 import { VariasiFindManyArgs } from "../../variasi/base/VariasiFindManyArgs";
 import { Variasi } from "../../variasi/base/Variasi";
 import { ProductService } from "../product.service";
-@common.UseGuards(GqlDefaultAuthGuard, gqlACGuard.GqlACGuard)
 @graphql.Resolver(() => Product)
 export class ProductResolverBase {
-  constructor(
-    protected readonly service: ProductService,
-    protected readonly rolesBuilder: nestAccessControl.RolesBuilder
-  ) {}
+  constructor(protected readonly service: ProductService) {}
 
-  @graphql.Query(() => MetaQueryPayload)
-  @nestAccessControl.UseRoles({
-    resource: "Product",
-    action: "read",
-    possession: "any",
-  })
   async _productsMeta(
     @graphql.Args() args: ProductCountArgs
   ): Promise<MetaQueryPayload> {
@@ -54,70 +38,46 @@ export class ProductResolverBase {
     };
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @graphql.Query(() => [Product])
-  @nestAccessControl.UseRoles({
-    resource: "Product",
-    action: "read",
-    possession: "any",
-  })
   async products(
     @graphql.Args() args: ProductFindManyArgs
   ): Promise<Product[]> {
-    return this.service.findMany(args);
+    return this.service.products(args);
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @graphql.Query(() => Product, { nullable: true })
-  @nestAccessControl.UseRoles({
-    resource: "Product",
-    action: "read",
-    possession: "own",
-  })
   async product(
     @graphql.Args() args: ProductFindUniqueArgs
   ): Promise<Product | null> {
-    const result = await this.service.findOne(args);
+    const result = await this.service.product(args);
     if (result === null) {
       return null;
     }
     return result;
   }
 
-  @common.UseInterceptors(AclValidateRequestInterceptor)
   @graphql.Mutation(() => Product)
-  @nestAccessControl.UseRoles({
-    resource: "Product",
-    action: "create",
-    possession: "any",
-  })
   async createProduct(
     @graphql.Args() args: CreateProductArgs
   ): Promise<Product> {
-    return await this.service.create({
+    return await this.service.createProduct({
       ...args,
       data: args.data,
     });
   }
 
-  @common.UseInterceptors(AclValidateRequestInterceptor)
   @graphql.Mutation(() => Product)
-  @nestAccessControl.UseRoles({
-    resource: "Product",
-    action: "update",
-    possession: "any",
-  })
   async updateProduct(
     @graphql.Args() args: UpdateProductArgs
   ): Promise<Product | null> {
     try {
-      return await this.service.update({
+      return await this.service.updateProduct({
         ...args,
         data: args.data,
       });
     } catch (error) {
       if (isRecordNotFoundError(error)) {
-        throw new apollo.ApolloError(
+        throw new GraphQLError(
           `No resource was found for ${JSON.stringify(args.where)}`
         );
       }
@@ -126,19 +86,14 @@ export class ProductResolverBase {
   }
 
   @graphql.Mutation(() => Product)
-  @nestAccessControl.UseRoles({
-    resource: "Product",
-    action: "delete",
-    possession: "any",
-  })
   async deleteProduct(
     @graphql.Args() args: DeleteProductArgs
   ): Promise<Product | null> {
     try {
-      return await this.service.delete(args);
+      return await this.service.deleteProduct(args);
     } catch (error) {
       if (isRecordNotFoundError(error)) {
-        throw new apollo.ApolloError(
+        throw new GraphQLError(
           `No resource was found for ${JSON.stringify(args.where)}`
         );
       }
@@ -146,14 +101,8 @@ export class ProductResolverBase {
     }
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @graphql.ResolveField(() => [Order], { name: "orders" })
-  @nestAccessControl.UseRoles({
-    resource: "Order",
-    action: "read",
-    possession: "any",
-  })
-  async resolveFieldOrders(
+  async findOrders(
     @graphql.Parent() parent: Product,
     @graphql.Args() args: OrderFindManyArgs
   ): Promise<Order[]> {
@@ -166,14 +115,8 @@ export class ProductResolverBase {
     return results;
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @graphql.ResolveField(() => [Variasi], { name: "variasi" })
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "read",
-    possession: "any",
-  })
-  async resolveFieldVariasi(
+  async findVariasi(
     @graphql.Parent() parent: Product,
     @graphql.Args() args: VariasiFindManyArgs
   ): Promise<Variasi[]> {
