@@ -1,29 +1,24 @@
-import { JwtService } from "@nestjs/jwt";
-import { mock } from "jest-mock-extended";
 import { TokenServiceBase } from "../../auth/base/token.service.base";
 import {
-  INVALID_PASSWORD_ERROR,
   INVALID_USERNAME_ERROR,
+  INVALID_PASSWORD_ERROR,
 } from "../../auth/constants";
 import { SIGN_TOKEN, VALID_CREDENTIALS, VALID_ID } from "./constants";
 
 describe("Testing the TokenServiceBase", () => {
   let tokenServiceBase: TokenServiceBase;
-  const jwtService = mock<JwtService>();
   beforeEach(() => {
-    tokenServiceBase = new TokenServiceBase(jwtService);
-    jwtService.signAsync.mockClear();
+    tokenServiceBase = new TokenServiceBase();
   });
   describe("Testing the BasicTokenService.createToken()", () => {
-    it("should create valid token for valid username and password", async () => {
-      jwtService.signAsync.mockReturnValue(Promise.resolve(SIGN_TOKEN));
+    it("should create valid token for given username and password", async () => {
       expect(
         await tokenServiceBase.createToken({
           id: VALID_ID,
-          username: VALID_CREDENTIALS.username,
-          password: VALID_CREDENTIALS.password,
+          username: "admin",
+          password: "admin",
         })
-      ).toBe(SIGN_TOKEN);
+      ).toBe("YWRtaW46YWRtaW4=");
     });
     it("should reject when username missing", () => {
       const result = tokenServiceBase.createToken({
