@@ -11,7 +11,13 @@ https://docs.amplication.com/how-to/custom-code
   */
 import { ObjectType, Field } from "@nestjs/graphql";
 import { ApiProperty } from "@nestjs/swagger";
-import { IsDate, IsString, IsOptional, ValidateNested } from "class-validator";
+import {
+  IsDate,
+  IsString,
+  MaxLength,
+  IsOptional,
+  ValidateNested,
+} from "class-validator";
 import { Type } from "class-transformer";
 import { Product } from "../../product/base/Product";
 
@@ -30,6 +36,7 @@ class Variasi {
     type: String,
   })
   @IsString()
+  @MaxLength(100000)
   @IsOptional()
   @Field(() => String, {
     nullable: true,
