@@ -16,41 +16,24 @@ import * as errors from "../../errors";
 import { Request } from "express";
 import { plainToClass } from "class-transformer";
 import { ApiNestedQuery } from "../../decorators/api-nested-query.decorator";
-import * as nestAccessControl from "nest-access-control";
-import * as defaultAuthGuard from "../../auth/defaultAuth.guard";
 import { VariasiService } from "../variasi.service";
-import { AclValidateRequestInterceptor } from "../../interceptors/aclValidateRequest.interceptor";
-import { AclFilterResponseInterceptor } from "../../interceptors/aclFilterResponse.interceptor";
 import { VariasiCreateInput } from "./VariasiCreateInput";
-import { VariasiWhereInput } from "./VariasiWhereInput";
-import { VariasiWhereUniqueInput } from "./VariasiWhereUniqueInput";
-import { VariasiFindManyArgs } from "./VariasiFindManyArgs";
-import { VariasiUpdateInput } from "./VariasiUpdateInput";
 import { Variasi } from "./Variasi";
+import { VariasiFindManyArgs } from "./VariasiFindManyArgs";
+import { VariasiWhereUniqueInput } from "./VariasiWhereUniqueInput";
+import { VariasiUpdateInput } from "./VariasiUpdateInput";
 import { ProductFindManyArgs } from "../../product/base/ProductFindManyArgs";
 import { Product } from "../../product/base/Product";
 import { ProductWhereUniqueInput } from "../../product/base/ProductWhereUniqueInput";
 
-@swagger.ApiBasicAuth()
-@common.UseGuards(defaultAuthGuard.DefaultAuthGuard, nestAccessControl.ACGuard)
 export class VariasiControllerBase {
-  constructor(
-    protected readonly service: VariasiService,
-    protected readonly rolesBuilder: nestAccessControl.RolesBuilder
-  ) {}
-  @common.UseInterceptors(AclValidateRequestInterceptor)
+  constructor(protected readonly service: VariasiService) {}
   @common.Post()
   @swagger.ApiCreatedResponse({ type: Variasi })
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "create",
-    possession: "any",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async create(@common.Body() data: VariasiCreateInput): Promise<Variasi> {
-    return await this.service.create({
+  async createVariasi(
+    @common.Body() data: VariasiCreateInput
+  ): Promise<Variasi> {
+    return await this.service.createVariasi({
       data: data,
       select: {
         createdAt: true,
@@ -61,21 +44,12 @@ export class VariasiControllerBase {
     });
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @common.Get()
   @swagger.ApiOkResponse({ type: [Variasi] })
   @ApiNestedQuery(VariasiFindManyArgs)
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "read",
-    possession: "any",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async findMany(@common.Req() request: Request): Promise<Variasi[]> {
+  async variasis(@common.Req() request: Request): Promise<Variasi[]> {
     const args = plainToClass(VariasiFindManyArgs, request.query);
-    return this.service.findMany({
+    return this.service.variasis({
       ...args,
       select: {
         createdAt: true,
@@ -86,22 +60,13 @@ export class VariasiControllerBase {
     });
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @common.Get("/:id")
   @swagger.ApiOkResponse({ type: Variasi })
   @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "read",
-    possession: "own",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async findOne(
+  async variasi(
     @common.Param() params: VariasiWhereUniqueInput
   ): Promise<Variasi | null> {
-    const result = await this.service.findOne({
+    const result = await this.service.variasi({
       where: params,
       select: {
         createdAt: true,
@@ -118,24 +83,15 @@ export class VariasiControllerBase {
     return result;
   }
 
-  @common.UseInterceptors(AclValidateRequestInterceptor)
   @common.Patch("/:id")
   @swagger.ApiOkResponse({ type: Variasi })
   @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "update",
-    possession: "any",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async update(
+  async updateVariasi(
     @common.Param() params: VariasiWhereUniqueInput,
     @common.Body() data: VariasiUpdateInput
   ): Promise<Variasi | null> {
     try {
-      return await this.service.update({
+      return await this.service.updateVariasi({
         where: params,
         data: data,
         select: {
@@ -158,19 +114,11 @@ export class VariasiControllerBase {
   @common.Delete("/:id")
   @swagger.ApiOkResponse({ type: Variasi })
   @swagger.ApiNotFoundResponse({ type: errors.NotFoundException })
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "delete",
-    possession: "any",
-  })
-  @swagger.ApiForbiddenResponse({
-    type: errors.ForbiddenException,
-  })
-  async delete(
+  async deleteVariasi(
     @common.Param() params: VariasiWhereUniqueInput
   ): Promise<Variasi | null> {
     try {
-      return await this.service.delete({
+      return await this.service.deleteVariasi({
         where: params,
         select: {
           createdAt: true,
@@ -189,15 +137,9 @@ export class VariasiControllerBase {
     }
   }
 
-  @common.UseInterceptors(AclFilterResponseInterceptor)
   @common.Get("/:id/products")
   @ApiNestedQuery(ProductFindManyArgs)
-  @nestAccessControl.UseRoles({
-    resource: "Product",
-    action: "read",
-    possession: "any",
-  })
-  async findManyProducts(
+  async findProducts(
     @common.Req() request: Request,
     @common.Param() params: VariasiWhereUniqueInput
   ): Promise<Product[]> {
@@ -222,11 +164,6 @@ export class VariasiControllerBase {
   }
 
   @common.Post("/:id/products")
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "update",
-    possession: "any",
-  })
   async connectProducts(
     @common.Param() params: VariasiWhereUniqueInput,
     @common.Body() body: ProductWhereUniqueInput[]
@@ -236,7 +173,7 @@ export class VariasiControllerBase {
         connect: body,
       },
     };
-    await this.service.update({
+    await this.service.updateVariasi({
       where: params,
       data,
       select: { id: true },
@@ -244,11 +181,6 @@ export class VariasiControllerBase {
   }
 
   @common.Patch("/:id/products")
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "update",
-    possession: "any",
-  })
   async updateProducts(
     @common.Param() params: VariasiWhereUniqueInput,
     @common.Body() body: ProductWhereUniqueInput[]
@@ -258,7 +190,7 @@ export class VariasiControllerBase {
         set: body,
       },
     };
-    await this.service.update({
+    await this.service.updateVariasi({
       where: params,
       data,
       select: { id: true },
@@ -266,11 +198,6 @@ export class VariasiControllerBase {
   }
 
   @common.Delete("/:id/products")
-  @nestAccessControl.UseRoles({
-    resource: "Variasi",
-    action: "update",
-    possession: "any",
-  })
   async disconnectProducts(
     @common.Param() params: VariasiWhereUniqueInput,
     @common.Body() body: ProductWhereUniqueInput[]
@@ -280,7 +207,7 @@ export class VariasiControllerBase {
         disconnect: body,
       },
     };
-    await this.service.update({
+    await this.service.updateVariasi({
       where: params,
       data,
       select: { id: true },
