@@ -10,10 +10,9 @@ https://docs.amplication.com/how-to/custom-code
 ------------------------------------------------------------------------------
   */
 import { Module } from "@nestjs/common";
-import { MorganModule } from "nest-morgan";
-import { ACLModule } from "../../auth/acl.module";
+
 @Module({
-  imports: [ACLModule, MorganModule],
-  exports: [ACLModule, MorganModule],
+  imports: [],
+  exports: [],
 })
 export class VariasiModuleBase {}
